Fixes to controls and car in editor
Due to how I used the input handler there was a bug where for example if you held a, then held w, then let go of a the game would think that you were still holding a rather than w. Essentially, it would only update when you let go and then pressed the other key. This is now fixed.

I also made it so that you can see the car in edit mode. It's just a temporary visual that get's disabled when the game is loaded, but it's nice to see it anyway.